# src/main/java/com/kietitmo/football_team_managerment/controllers/RoleController.java
package com.kietitmo.football_team_managerment.controllers;

import com.kietitmo.football_team_managerment.dto.request.RoleRequest;
import com.kietitmo.football_team_managerment.dto.response.ApiResponse;
import com.kietitmo.football_team_managerment.dto.response.PageResponse;
import com.kietitmo.football_team_managerment.dto.response.RoleResponse;
import com.kietitmo.football_team_managerment.services.RoleService;
import lombok.AccessLevel;
import lombok.RequiredArgsConstructor;
import lombok.experimental.FieldDefaults;
import lombok.extern.slf4j.Slf4j;
import org.springframework.data.domain.Pageable;
import org.springframework.web.bind.annotation.*;

import java.util.List;

@RestController
@RequestMapping("/roles")
@RequiredArgsConstructor
@FieldDefaults(level = AccessLevel.PRIVATE, makeFinal = true)
@Slf4j
public class RoleController {
    RoleService roleService;

    @PostMapping
    ApiResponse<RoleResponse> create(@RequestBody RoleRequest request){
        return ApiResponse.<RoleResponse>builder()
                .result(roleService.create(request))
                .build();
    }

    @GetMapping
    ApiResponse<List<RoleResponse>> getAll(){
        return ApiResponse.<List<RoleResponse>>builder()
                .result(roleService.getAll())
                .build();
    }

    @PutMapping("/{role}")
    ApiResponse<RoleResponse> update(@PathVariable String role, @RequestBody RoleRequest request){

        return ApiResponse.<RoleResponse>builder()
                .result(roleService.updateRole(role, request))
                .build();
    }

    @DeleteMapping("/{role}")
    ApiResponse<Void> delete(@PathVariable String role){
        roleService.delete(role);
        return ApiResponse.<Void>builder().build();
    }

    @GetMapping(path = "/advance-search-with-specification")
    public ApiResponse<PageResponse<List<RoleResponse>>> advanceSearchWithSpecifications(Pageable pageable,
                                                                                             @RequestParam(required = false) String[] role) {
        return ApiResponse.<PageResponse<List<RoleResponse>>>builder()
                .result(roleService.advanceSearchWithSpecifications(pageable, role))
                .build();
    }
}