# src/main/resources/api-docs.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Football team management",
    "version": "3.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8888",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "SecurityScheme": []
    }
  ],
  "paths": {
...
    "/roles/{role}": {
      "put": {
        "tags": [
          "role-controller"
        ],
        "operationId": "update_3",
        "parameters": [
          {
            "name": "role",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseRoleResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "role-controller"
        ],
        "operationId": "delete_3",
        "parameters": [
          {
            "name": "role",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          }
        }
      }
    },
    "/matches/{match}": {
      "put": {
        "tags": [
          "match-controller"
        ],
        "operationId": "update_4",
        "parameters": [
          {
            "name": "match",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseMatchResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "match-controller"
        ],
        "operationId": "delete_5",
        "parameters": [
          {
            "name": "match",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          }
        }
      }
    },
...
    "/roles": {
      "get": {
        "tags": [
          "role-controller"
        ],
        "operationId": "getAll_3",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseListRoleResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "role-controller"
        ],
        "operationId": "create_3",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseRoleResponse"
                }
              }
            }
          }
        }
      }
    },
    "/permissions": {
      "get": {
        "tags": [
          "permission-controller"
        ],
        "operationId": "getAll_4",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseListPermissionResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "permission-controller"
        ],
        "operationId": "create_4",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PermissionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePermissionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/matches": {
      "get": {
        "tags": [
          "match-controller"
        ],
        "operationId": "getAll_5",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseListMatchResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "match-controller"
        ],
        "operationId": "create_5",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseMatchResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/votes/advance-search-with-specification": {
      "get": {
        "tags": [
          "vote-controller"
        ],
        "operationId": "advanceSearchWithSpecifications",
        "parameters": [
          {
            "name": "pageable",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          },
          {
            "name": "vote",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePageResponseListVoteResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/users/advance-search-with-specification": {
      "get": {
        "tags": [
          "user-controller"
        ],
        "operationId": "advanceSearchWithSpecifications_1",
        "parameters": [
          {
            "name": "pageable",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          },
          {
            "name": "user",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePageResponseListUserResponse"
                }
              }
            }
          }
        }
      }
    },
    "/trainings/advance-search-with-specification": {
      "get": {
        "tags": [
          "training-controller"
        ],
        "operationId": "advanceSearchWithSpecifications_2",
        "parameters": [
          {
            "name": "pageable",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          },
          {
            "name": "training",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePageResponseListTrainingResponse"
                }
              }
            }
          }
        }
      }
    },
    "/teams/advance-search-with-specification": {
      "get": {
        "tags": [
          "team-controller"
        ],
        "operationId": "advanceSearchWithSpecifications_3",
        "parameters": [
          {
            "name": "pageable",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          },
          {
            "name": "team",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePageResponseListTeamResponse"
                }
              }
            }
          }
        }
      }
    },
    "/roles/advance-search-with-specification": {
      "get": {
        "tags": [
          "role-controller"
        ],
        "operationId": "advanceSearchWithSpecifications_4",
        "parameters": [
          {
            "name": "pageable",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          },
          {
            "name": "role",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePageResponseListRoleResponse"
                }
              }
            }
          }
        }
      }
    },
    "/permissions/advance-search-with-specification": {
      "get": {
        "tags": [
          "permission-controller"
        ],
        "operationId": "advanceSearchWithSpecifications_5",
        "parameters": [
          {
            "name": "pageable",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          },
          {
            "name": "permission",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePageResponseListPermissionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/matches/advance-search-with-specification": {
      "get": {
        "tags": [
          "match-controller"
        ],
        "operationId": "advanceSearchWithSpecifications_6",
        "parameters": [
          {
            "name": "pageable",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          },
          {
            "name": "match",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePageResponseListMatchResponse"
                }
              }
            }
          }
        }
      }
    },
    "/eventParticipants/advance-search-with-specification": {
      "get": {
        "tags": [
          "event-participant-controller"
        ],
        "operationId": "advanceSearchWithSpecifications_7",
        "parameters": [
          {
            "name": "pageable",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          },
          {
            "name": "participant",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePageResponseListEventParticipantResponse"
                }
              }
            }
          }
        }
      }
    },
    "/contents/advance-search-with-specification": {
      "get": {
        "tags": [
          "content-controller"
        ],
        "operationId": "advanceSearchWithSpecifications_8",
        "parameters": [
          {
            "name": "pageable",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          },
          {
            "name": "content",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePageResponseListContentResponse"
                }
              }
            }
          }
        }
      }
    },
    "/permissions/{permission}": {
      "delete": {
        "tags": [
          "permission-controller"
        ],
        "operationId": "delete_4",
        "parameters": [
          {
            "name": "permission",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "EventParticipantRequest": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          },
          "eventId": {
            "type": "string"
          },
          "roleInEvent": {
            "type": "string",
            "enum": [
              "COACH",
              "CAPTAIN",
              "PLAYER",
              "MEDICINE_TEAM",
              "OTHER"
            ]
          }
        }
      },
...
      "RoleRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "permissions": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "ApiResponseRoleResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "result": {
            "$ref": "#/components/schemas/RoleResponse"
          }
        }
      },
      "PermissionResponse": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
      "RoleResponse": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "permissions": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PermissionResponse"
            }
          }
        }
      },
      "MatchRequest": {
        "type": "object",
        "properties": {
          "homeTeam": {
            "type": "string"
          },
          "awayTeam": {
            "type": "string"
          },
          "matchDate": {
            "type": "string",
            "format": "date"
          },
          "homeGoals": {
            "type": "integer",
            "format": "int32"
          },
          "awayGoals": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "participants": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EventParticipantRequest"
            }
          },
          "creationDate": {
            "type": "string",
            "format": "date"
          },
          "lastUpdateDate": {
            "type": "string",
            "format": "date"
          }
        }
      },
      "ApiResponseMatchResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "result": {
            "$ref": "#/components/schemas/MatchResponse"
          }
        }
      },
      "MatchResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "homeTeamId": {
            "type": "string"
          },
          "awayTeamId": {
            "type": "string"
          },
          "matchDate": {
            "type": "string",
            "format": "date"
          },
          "homeGoals": {
            "type": "integer",
            "format": "int32"
          },
          "awayGoals": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "participants": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EventParticipantResponse"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "PermissionRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
      "ApiResponsePermissionResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "result": {
            "$ref": "#/components/schemas/PermissionResponse"
          }
        }
      },
...
      "ApiResponseVoid": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "result": {
            "type": "object"
          }
        }
      },
...
      "Pageable": {
        "type": "object",
        "properties": {
          "page": {
            "minimum": 0,
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "ApiResponsePageResponseListVoteResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "result": {
            "$ref": "#/components/schemas/PageResponseListVoteResponse"
          }
        }
      },
      "PageResponseListVoteResponse": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalPage": {
            "type": "integer",
            "format": "int64"
          },
          "totalItem": {
            "type": "integer",
            "format": "int64"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VoteResponse"
            }
          }
        }
      },
...
      "ApiResponsePageResponseListUserResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "result": {
            "$ref": "#/components/schemas/PageResponseListUserResponse"
          }
        }
      },
      "PageResponseListUserResponse": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalPage": {
            "type": "integer",
            "format": "int64"
          },
          "totalItem": {
            "type": "integer",
            "format": "int64"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserResponse"
            }
          }
        }
      },
...
      "ApiResponsePageResponseListTrainingResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "result": {
            "$ref": "#/components/schemas/PageResponseListTrainingResponse"
          }
        }
      },
      "PageResponseListTrainingResponse": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalPage": {
            "type": "integer",
            "format": "int64"
          },
          "totalItem": {
            "type": "integer",
            "format": "int64"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TrainingResponse"
            }
          }
        }
      },
...
      "ApiResponsePageResponseListTeamResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "result": {
            "$ref": "#/components/schemas/PageResponseListTeamResponse"
          }
        }
      },
...
      "ApiResponseListRoleResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "result": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoleResponse"
            }
          }
        }
      },
      "ApiResponsePageResponseListRoleResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "result": {
            "$ref": "#/components/schemas/PageResponseListRoleResponse"
          }
        }
      },
...
      "ApiResponseListPermissionResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "result": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PermissionResponse"
            }
          }
        }
      },
      "ApiResponsePageResponseListPermissionResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "result": {
            "$ref": "#/components/schemas/PageResponseListPermissionResponse"
          }
        }
      },
...
      "ApiResponseListMatchResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "result": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MatchResponse"
            }
          }
        }
      },
      "ApiResponsePageResponseListMatchResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "result": {
            "$ref": "#/components/schemas/PageResponseListMatchResponse"
          }
        }
      },
...
      "ApiResponsePageResponseListEventParticipantResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "result": {
            "$ref": "#/components/schemas/PageResponseListEventParticipantResponse"
          }
        }
      },
...
      "ApiResponsePageResponseListContentResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "result": {
            "$ref": "#/components/schemas/PageResponseListContentResponse"
          }
        }
      },
...
    },
    "securitySchemes": {
      "SecurityScheme": {
        "type": "http",
        "name": "SecurityScheme",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}